Split bot/console.py into multiple parts
